Assignment 2/requirements_enhanced.txt
requests>=2.25.1
python-dotenv>=0.19.0
PyJWT>=2.0.0
aiohttp>=3.8.0
